+++ packages/viz/src/components/trends/cbg/CBGSliceAnimated.js
@@ -19,12 +19,12 @@ import _ from 'lodash'
 import cx from 'classnames'
 import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
-import { TransitionMotion, spring } from '@serprex/react-motion'
+import { spring, TransitionMotion } from '@serprex/react-motion'
 
 import { springConfig } from '../../../utils/constants'
 import withDefaultYPosition from '../common/withDefaultYPosition'
 
-import CBGSliceSegment from './CBGSliceSegment'
+import { CbgSliceSegment } from 'dumb'
 
 import styles from './CBGSliceAnimated.css'
 
@@ -205,24 +205,26 @@ export class CBGSliceAnimated extends PureComponent {
           return (
             <g id={`cbgSlice-${datum.id}`}>
               {_.map(interpolateds, (interpolated) => {
-                const segment = renderPieces[interpolated.key]
+                const key = interpolated.key
+                const segment = renderPieces[key]
                 const classes = cx({
                   [styles.segment]: true,
                   [styles[segment.classKey]]: !showingCbgDateTraces,
                   [styles[`${segment.classKey}Faded`]]: showingCbgDateTraces
                 })
                 return (
-                  <CBGSliceSegment
+                  <CbgSliceSegment
                     classes={classes}
                     datum={datum}
-                    interpolated={interpolated}
-                    key={interpolated.key}
+                    id={key}
+                    key={key}
                     positionData={{
                       left: xScale(datum.msX),
                       tooltipLeft: datum.msX > tooltipLeftThreshold,
                       yPositions
                     }}
                     segment={segment}
+                    style={interpolated.style}
                     width={width}
                     x={binLeftX}
                   />